// src/common/types/generated/types.ts
export type Maybe<T> = T | null
export type InputMaybe<T> = Maybe<T>
// ...
export type Scalars = {
  ID: { input: string | number; output: string }
  String: { input: string; output: string }
  Boolean: { input: boolean; output: boolean }
  Int: { input: number; output: number }
  Float: { input: number; output: number }
  DateTime: { input: any; output: any }
  Upload: { input: any; output: any }
}

export type IBoard = {
  __typename?: 'Board'
  _id: Scalars['ID']['output']
  boardAddress?: Maybe<IBoardAddress>
  contents: Scalars['String']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  dislikeCount: Scalars['Int']['output']
  images?: Maybe<Array<Scalars['String']['output']>>
  likeCount: Scalars['Int']['output']
  title: Scalars['String']['output']
  updatedAt: Scalars['DateTime']['output']
  user?: Maybe<IUser>
  writer?: Maybe<Scalars['String']['output']>
  youtubeUrl?: Maybe<Scalars['String']['output']>
}

export type IBoardAddress = {
  __typename?: 'BoardAddress'
  _id: Scalars['ID']['output']
  address?: Maybe<Scalars['String']['output']>
  addressDetail?: Maybe<Scalars['String']['output']>
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  updatedAt: Scalars['DateTime']['output']
  zipcode?: Maybe<Scalars['String']['output']>
}

export type IBoardAddressInput = {
  address?: InputMaybe<Scalars['String']['input']>
  addressDetail?: InputMaybe<Scalars['String']['input']>
  zipcode?: InputMaybe<Scalars['String']['input']>
}

export type IBoardComment = {
  __typename?: 'BoardComment'
  _id: Scalars['ID']['output']
  contents: Scalars['String']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  rating: Scalars['Float']['output']
  updatedAt: Scalars['DateTime']['output']
  user?: Maybe<IUser>
  writer?: Maybe<Scalars['String']['output']>
}
// ...
export type ICreateUseditemInput = {
  contents: Scalars['String']['input']
  images?: InputMaybe<Array<Scalars['String']['input']>>
  name: Scalars['String']['input']
  price: Scalars['Int']['input']
  remarks: Scalars['String']['input']
  tags?: InputMaybe<Array<Scalars['String']['input']>>
  useditemAddress?: InputMaybe<IUseditemAddressInput>
}

export type ICreateUseditemQuestionAnswerInput = {
  contents: Scalars['String']['input']
}

export type ICreateUseditemQuestionInput = {
  contents: Scalars['String']['input']
}

export type ICreateUserInput = {
  email: Scalars['String']['input']
  name: Scalars['String']['input']
  password: Scalars['String']['input']
}

export type IFileManager = {
  __typename?: 'FileManager'
  _id: Scalars['ID']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  isUsed: Scalars['Boolean']['output']
  size?: Maybe<Scalars['Float']['output']>
  updatedAt: Scalars['DateTime']['output']
  url: Scalars['String']['output']
}

export type IMutation = {
  __typename?: 'Mutation'
  createBoard: IBoard
  createBoardComment: IBoardComment
  createPointTransactionOfBuyingAndSelling: IUseditem
  createPointTransactionOfLoading: IPointTransaction
  createUseditem: IUseditem
  createUseditemQuestion: IUseditemQuestion
  createUseditemQuestionAnswer: IUseditemQuestionAnswer
  createUser: IUser
  deleteBoard: Scalars['ID']['output']
  deleteBoardComment: Scalars['ID']['output']
  deleteBoards: Array<Scalars['ID']['output']>
  deleteUseditem: Scalars['ID']['output']
  deleteUseditemQuestion: Scalars['ID']['output']
  deleteUseditemQuestionAnswer: Scalars['String']['output']
  dislikeBoard: Scalars['Int']['output']
  likeBoard: Scalars['Int']['output']
  loginUser: IToken
  loginUserExample: IToken
  logoutUser: Scalars['Boolean']['output']
  resetUserPassword: Scalars['Boolean']['output']
  restoreAccessToken: IToken
  toggleUseditemPick: Scalars['Int']['output']
  updateBoard: IBoard
  updateBoardComment: IBoardComment
  updateUseditem: IUseditem
  updateUseditemQuestion: IUseditemQuestion
  updateUseditemQuestionAnswer: IUseditemQuestionAnswer
  updateUser: IUser
  uploadFile: IFileManager
}
// ...
export type IMutationCreateUseditemArgs = {
  createUseditemInput: ICreateUseditemInput
}
// ...
export type IMutationUploadFileArgs = {
  file: Scalars['Upload']['input']
}

export type IPointTransaction = {
  __typename?: 'PointTransaction'
  _id: Scalars['ID']['output']
  amount: Scalars['Int']['output']
  balance: Scalars['Int']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  impUid?: Maybe<Scalars['ID']['output']>
  status: Scalars['String']['output']
  statusDetail: Scalars['String']['output']
  updatedAt: Scalars['DateTime']['output']
  useditem?: Maybe<IUseditem>
  user?: Maybe<IUser>
}
// ...
export type IToken = {
  __typename?: 'Token'
  accessToken: Scalars['String']['output']
}
// ...
export type IUseditem = {
  __typename?: 'Useditem'
  _id: Scalars['ID']['output']
  buyer?: Maybe<IUser>
  contents: Scalars['String']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  images?: Maybe<Array<Scalars['String']['output']>>
  name: Scalars['String']['output']
  pickedCount?: Maybe<Scalars['Int']['output']>
  price?: Maybe<Scalars['Int']['output']>
  remarks: Scalars['String']['output']
  seller?: Maybe<IUser>
  soldAt?: Maybe<Scalars['DateTime']['output']>
  tags?: Maybe<Array<Scalars['String']['output']>>
  updatedAt: Scalars['DateTime']['output']
  useditemAddress?: Maybe<IUseditemAddress>
}

export type IUseditemAddress = {
  __typename?: 'UseditemAddress'
  _id: Scalars['ID']['output']
  address?: Maybe<Scalars['String']['output']>
  addressDetail?: Maybe<Scalars['String']['output']>
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  lat?: Maybe<Scalars['Float']['output']>
  lng?: Maybe<Scalars['Float']['output']>
  updatedAt: Scalars['DateTime']['output']
  zipcode?: Maybe<Scalars['String']['output']>
}

export type IUseditemAddressInput = {
  address?: InputMaybe<Scalars['String']['input']>
  addressDetail?: InputMaybe<Scalars['String']['input']>
  lat?: InputMaybe<Scalars['Float']['input']>
  lng?: InputMaybe<Scalars['Float']['input']>
  zipcode?: InputMaybe<Scalars['String']['input']>
}

export type IUseditemQuestion = {
  __typename?: 'UseditemQuestion'
  _id: Scalars['ID']['output']
  contents: Scalars['String']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  updatedAt: Scalars['DateTime']['output']
  useditem: IUseditem
  user: IUser
}

export type IUseditemQuestionAnswer = {
  __typename?: 'UseditemQuestionAnswer'
  _id: Scalars['ID']['output']
  contents: Scalars['String']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  updatedAt: Scalars['DateTime']['output']
  useditemQuestion: IUseditemQuestion
  user: IUser
}

export type IUser = {
  __typename?: 'User'
  _id: Scalars['ID']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  email: Scalars['String']['output']
  name: Scalars['String']['output']
  picture?: Maybe<Scalars['String']['output']>
  updatedAt: Scalars['DateTime']['output']
  userPoint?: Maybe<IUserPoint>
}

export type IUserPoint = {
  __typename?: 'UserPoint'
  _id: Scalars['ID']['output']
  amount: Scalars['Int']['output']
  createdAt: Scalars['DateTime']['output']
  deletedAt?: Maybe<Scalars['DateTime']['output']>
  updatedAt: Scalars['DateTime']['output']
  user: IUser
}